wip: connect did select
