--- src/features/connect/ConnectionRequestScreen.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import { SafeAreaView, Text } from "react-native";
-
-const ConnectionRequestScreen = () => {
-  return (
-    <SafeAreaView>
-      <Text>Hello, ConnectionRequestScreen!</Text>
-    </SafeAreaView>
-  );
-};
-
-export default ConnectionRequestScreen;
