+++ src/features/connect/ReviewConnectionScreen.tsx
@@ -1,12 +1,12 @@
+import React from "react";
+import { View, SafeAreaView, Text, StyleSheet, ScrollView } from "react-native";
 import { Button } from "@/components/Button";
 import { BadgeNames, Item } from "@/components/Item";
 import { LabelValueItem } from "@/components/LabelValue";
 import { ColorTheme } from "@/theme/colors";
-import { FlexLayouts, Layouts } from "@/theme/layouts";
+import { SPACE } from "@/theme/layouts";
 import { Typography } from "@/theme/typography";
 import { formatDID, formatDate } from "@/util/formatters";
-import React from "react";
-import { View, SafeAreaView, Text } from "react-native";
 
 const ReviewConnectionScreen = () => {
   const { connection, profile } = {
@@ -23,66 +23,76 @@ const ReviewConnectionScreen = () => {
   };
 
   return (
-    <SafeAreaView>
-      <View style={Layouts.container}>
-        <View style={Layouts.row}>
-          <Item
-            heading={profile.name}
-            subtitle={profile.displayName}
-            body={formatDID(profile.id)}
-            iconName={profile.icon}
-            badgeName={BadgeNames.PROFILE}
-            headingSize="heading3"
-          />
-          <Item
-            heading={connection.name}
-            iconName={connection.icon}
-            badgeName={BadgeNames.CONNECTION}
-            headingSize="heading3"
-          />
+    <SafeAreaView style={styles.wrapper}>
+      <ScrollView>
+        <View style={styles.container}>
+          <View style={styles.row}>
+            <Item
+              heading={profile.name}
+              subtitle={profile.displayName}
+              body={formatDID(profile.id)}
+              iconName={profile.icon}
+              badgeName={BadgeNames.PROFILE}
+              headingSize="heading4"
+            />
+            <Item
+              heading={connection.name}
+              iconName={connection.icon}
+              badgeName={BadgeNames.CONNECTION}
+              headingSize="heading4"
+            />
+          </View>
+          <View style={styles.row}>
+            <Text style={Typography.body4}>
+              {connection.name} is connected to Professional profile and has
+              access to certain info.
+            </Text>
+          </View>
+          <View>
+            <LabelValueItem
+              label="Connection made"
+              value={formatDate(new Date().toString())}
+            />
+          </View>
+          <View style={styles.column}>
+            <Text style={Typography.body2}>{connection.name} can</Text>
+            <Text style={Typography.body4}>
+              &bull; see and edit your profile info
+            </Text>
+            <Text style={Typography.body4}>
+              &bull; see and edit your contacts
+            </Text>
+            <Text style={Typography.body4}>&bull; see and edit your chats</Text>
+          </View>
+          <View style={styles.column}>
+            <Button
+              kind="destructive"
+              text={`Disconnect ${connection.name} from ${profile.name}`}
+            />
+            <Text style={disclaimerText}>
+              Disconnecting may take up to 24 hours
+            </Text>
+          </View>
         </View>
-        <View style={Layouts.row}>
-          <Text style={Typography.body4}>
-            <Text style={Typography.body2}>{connection.name}</Text> is connected
-            to Professional profile and has access to certain info.
-          </Text>
-        </View>
-        <View>
-          <LabelValueItem
-            label="Connection made"
-            value={formatDate(new Date().toString())}
-          />
-        </View>
-        <View style={Layouts.row}>
-          <Text style={Typography.body4}>
-            <Text style={Typography.body2}>{connection.name}</Text> can
-          </Text>
-          <Text style={Typography.body4}>
-            &bull; see and edit your profile info
-          </Text>
-          <Text style={Typography.body4}>
-            &bull; see and edit your contacts
-          </Text>
-          <Text style={Typography.body4}>&bull; see and edit your chats</Text>
-        </View>
-        <View style={FlexLayouts.column}>
-          <Button
-            kind="destructive"
-            text={`Disconnect ${connection.name} from ${profile.name}`}
-          />
-          <Text
-            style={[
-              Typography.label3,
-              Typography.textCenter,
-              { color: ColorTheme.REDUCED },
-            ]}
-          >
-            Disconnecting may take up to 24 hours
-          </Text>
-        </View>
-      </View>
+      </ScrollView>
     </SafeAreaView>
   );
 };
+
+const styles = StyleSheet.create({
+  wrapper: { flex: 1 },
+  container: {
+    padding: SPACE.BASE,
+    gap: SPACE.MEDIUM,
+  },
+  column: { gap: SPACE.XSMALL },
+  row: { gap: SPACE.SMALL },
+});
+
+const disclaimerText = StyleSheet.create([
+  Typography.label3,
+  Typography.textCenter,
+  { color: ColorTheme.REDUCED },
+]);
 
 export default ReviewConnectionScreen;
