--- src/features/connect/SelectProfilesScreen.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { View, SafeAreaView } from "react-native";
-import { Layouts } from "@/theme/layouts";
-
-const ReviewConnectionScreen = () => {
-  return (
-    <SafeAreaView>
-      <View style={Layouts.container}></View>
-    </SafeAreaView>
-  );
-};
-
-export default ReviewConnectionScreen;
